+++ frontend/src/components/search_customer/rs/page.tsx
@@ -177,11 +177,6 @@ const SearchCustomerRepairshoprScreen = () => {
         setSelectedCustomer(null);
     };
 
-
-
-
-
-
     return (
         <main className="flex justify-center items-center h-screen bg-orange-100">