+++ server/controllers/customer_visits/create_customer_visit.js
@@ -5,7 +5,7 @@ const addCustomerSchema = Yup.object({
     firstname: Yup.string().required("Firstname is required!"),
     lastname: Yup.string().required("Lastname is required!"),
     businessname: Yup.string(),
-    email: Yup.string().required("Email is required!"),
+    email: Yup.string(),
     phone: Yup.string(),
     mobile: Yup.string(),
     address: Yup.string(),
@@ -72,13 +72,11 @@ const addCustomer = async (req, res) => {
             "INSERT INTO customer_visits (customer_id, visit_date) VALUES ($1, $2)",
             [customerId, visit_date]
         );
-        console.log("customerId, visit_date", customerId + visit_date);
         return res.status(201).json({
             message:
                 "Successfully added as visit\nBooking agent will now book you in",
         });
     } catch (error) {
-        console.log(error);
         // Handle validation or other errors
         const errors = {};
         if (error.inner) {