+++ server/controllers/department/hhp/technicians/add_hhp_task.js
@@ -8,6 +8,7 @@ const AddHHPTaskSchema = Yup.object({
     model: Yup.string().required("Model is required"),
     warranty: Yup.string().required("Warranty is required"),
     engineer: Yup.string(),
+    repairshopr_customer_id: Yup.string(),
     fault: Yup.string().required("Fault is required"),
     imei: Yup.string().required("IMEI is required"),
     serial_number: Yup.string().required("Serial number is required"),
@@ -42,6 +43,7 @@ const AddHHPTask = async (req, res) => {
         repeat_repair,
         created_at,
         additional_info,
+        repairshopr_customer_id,
     } = req.body;
 
     try {
@@ -58,7 +60,7 @@ const AddHHPTask = async (req, res) => {
             });
         } else {
             const { rows } = await pool.query(
-                "INSERT INTO technician_tasks (service_order_no, date_booked, model, warranty, engineer, fault, imei, serial_number, unit_status, ticket_number, department, job_added_by, stores, repairshopr_job_id, repeat_repair, created_at, additional_info) values ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15, $16, $17) returning id",
+                "INSERT INTO technician_tasks (service_order_no, date_booked, model, warranty, engineer, fault, imei, serial_number, unit_status, ticket_number, department, job_added_by, stores, repairshopr_job_id, repeat_repair, created_at, additional_info, repairshopr_customer_id) values ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15, $16, $17, $18) returning id",
                 [
                     service_order_no,
                     date_booked,
@@ -77,6 +79,7 @@ const AddHHPTask = async (req, res) => {
                     repeat_repair,
                     created_at,
                     additional_info,
+                    repairshopr_customer_id,
                 ]
             );
             const fetchResult = await pool.query(returnDataWithNewRow, [