+++ server/server.js
@@ -28,6 +28,7 @@ import { router as cars } from "./routes/cars/index.js";
 import { router as fuel } from "./routes/fuel_consumption/index.js";
 import { router as devices } from "./routes/devices/index.js";
 import { router as customer_visits } from "./routes/customer_visits/index.js";
+import { router as customers } from "./routes/customers/index.js";
 
 const { router: fileRoutes } = QCfileRoutesModule;
 const { router: dtvFiles } = dtvhaFiles;
@@ -71,6 +72,7 @@ app.use("/api/v1/comments", ticketComments);
 app.use("/api/v1/customer_visits", customer_visits);
 app.use("/api/v1/fuel", fuel);
 app.use("/api/v1/devices", devices);
+app.use("/api/v1/customers", customers);
 app.use("/tools/dev/add_images_checklist", add_images_checklist);
 
 const PORT = process.env.NEXT_PUBLIC_EXPRESS_SERVER_PORT;